쿼타북 데이터 타입 에러 수정 (#592)

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/dto/external/quotabook/StakeholdersRes.java
@@ -48,17 +48,17 @@ public class StakeholdersRes {
 
         List<Stakeholder> top5 = new ArrayList<>(stakeholders.subList(0, TOP_N - 1));
         long etcNumShares = 0L;
-        long etcOwnerShipRatio = 0L;
+        float etcOwnerShipRatio = 0;
 
         for(Stakeholder holder: stakeholders.subList(TOP_N, stakeholders.size()-1)) {
             etcNumShares += holder.numShares;
-            etcOwnerShipRatio += Long.valueOf(holder.ownership);
+            etcOwnerShipRatio += Float.valueOf(holder.ownership);
         }
 
         Stakeholder etcStakeholder = new Stakeholder();
         etcStakeholder.setStakeholderName("기타주주");
         etcStakeholder.setNumShares(etcNumShares);
-        etcStakeholder.setOwnership(Long.toString(etcOwnerShipRatio));
+        etcStakeholder.setOwnership(Float.toString(etcOwnerShipRatio));
 
         top5.add(etcStakeholder);
